Moved a lot of the stuff from the BallControl project into the ArmadilloZ project.  Added a bunch more animations.
All of the walking animation directions should be in there now as well as the idle directions.  The rolling directions are also done except for the ones that we don't have the art for yet (back left, back right, straight back and straight forward).  I added in the slam animation too, but didn't hook it up because I wasn't sure how we wanted the attack animations done.